pyproject.toml
[tool.poetry]
name = "prod"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.3"
bcrypt = ">=4.2.0,<4.3.0"
pyjwt = ">=2.9.0,<2.10.0"
fastapi = ">=0.115.0,<0.116.0"
sqlalchemy = ">=2.0.35,<2.1.0"
pydantic = {extras = ["email"], version = "^2.9.2"}
uvicorn = ">=0.31.0,<0.32.0"
alembic = "^1.13.3"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
